Refactor: simplify suite/run_test/context codependence
It's easier and, frankly, clearer, if we just put it all into the
`sisyphus` constructor. Separating things makes it less clear what is
passed in where, and what belongs in context and what doesn't, and where
context should be created or modified.

Everything goes in context. Context is created per-suite. Every
reference to run_test, suite, length, reporter, etc. -- anything in
context -- must be directly accessed from context. The test-runner is
created at the same time as context, and their circular dependence is
resolved during context initialization. Suite is no longer responsible
for correctly associating the test_run function with context, since it
is meant to just be a part of context from the perspective of the suite.

With any luck, careful handling of shared context will make
parallelization far easier in the near future.