Refactor/Major: suite: object definition shorthand
A suite is a description, and an array or object that defines tests.
Previously, the object and array were treated more-or-less the same:
object keys just give descriptions to tests which otherwise would only
be addressable by their index. Now, object definition is a shorthand for
defining an array of sub-suites. Internally, a suite definition is
always an array of tests, and tests no longer have descriptions, only
indices; a "test with a description" is a sub-suite of one test.