Refactor: standardize reporter APIs: send/receive to and from reporters
The same flow exists in Sisyphus' suite contexts as does in method
interceptors, but it looks and sounds different. That's silly. This
commit makes it all consistent.

The flow is: a message is originated. The originator of the message
`send`s the message to a dispatcher. The dispatcher gives the message to
a `reporter`, and then `receive`s a reporter and a report in response.
The report on the message is then forwarded to the orginator. The
reporter that returns with a report in `receive` may not be the same
reporter to which the report was sent; so, repeated calls to `send` may
end up dispatching messages to different reporters, depending on what
`reporter` is `receive`d in response. That is, the "next" reporter is
always the one that returns with the latest report, which may not be the
same one reporter to which the last-sent message was given. (In other
words, reporters can hand off work to one another; they are mutually
recursive.)

Previously, Sisyphus suite contexts performed this flow with `report`
and `next_reporter` on `reporter`s. Interceptors performed this flow
with `intercept` (which did the equivalent of `send` and `receive`)
on `receiver`s. Now, they both perform this flow using `send` and
`receive` on `reporter`s.